// src/shared/generated/WorkspaceCreateInputRecord.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { WorkspaceSlotInputRecord } from "./WorkspaceSlotInputRecord";

export type WorkspaceCreateInputRecord = { name: string, template?: string, browserZoomMode?: string, browserZoomPercent?: number, slots?: Array<WorkspaceSlotInputRecord>, };